Login and profile cards refactored
