+++ lib/pages/chat/chat_room_list_item.dart
@@ -1,6 +1,3 @@
-import 'dart:async';
-
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:eatwithme/models/models.dart';
 import 'package:eatwithme/pages/chat/chat_room.dart';
 import 'package:eatwithme/pages/profile/profile.dart';
