+++ lib/pages/profile/editProfile.dart
@@ -12,7 +12,7 @@ import 'package:flutter/material.dart';
 import 'package:eatwithme/pages/interests/interests.dart';
 import 'package:eatwithme/main.dart';
 import 'package:eatwithme/widgets/loadingCircle.dart';
-import 'package:eatwithme/pages/auth/auth.dart';
+import 'package:eatwithme/services/auth.dart';
 import 'package:image_picker/image_picker.dart';
 import 'package:firebase_storage/firebase_storage.dart';
 
@@ -56,7 +56,7 @@ class _ProfilePageState extends State<EditProfilePage> {
     setState(() {
       Firestore.instance
           .collection('Users')
-          .document(authService.currentUid)
+          .document(widget.uid)
           .setData(
         {
           'photoURL': downloadURL,
@@ -105,172 +105,173 @@ class _ProfilePageState extends State<EditProfilePage> {
 
   @override
   Widget build(BuildContext context) {
+    return Container();
     // Set initial values on first load
-    return SafeArea(
-      child: ListView(
-        children: <Widget>[
-          Container(
-            child: Stack(
-              alignment: Alignment.topCenter,
-              children: <Widget>[
-                Column(
-                  children: <Widget>[
-                    Container(
-                      height: 150.0,
-                    ),
-                    Card(
-                        child: Column(
-                      mainAxisSize: MainAxisSize.min,
-                      children: <Widget>[
-                        Container(
-                          alignment: Alignment(0.0, 1.0),
-                          height: 150.0,
-                          width: 500.0,
-                          color: Color(0x00333333),
-                          child: TextField(
-                            decoration: InputDecoration(
-                                prefixIcon: Icon(Icons.account_circle),
-                                helperText: "Your display name"),
-                            style: TextStyle(
-                                color: Colors.black,
-                                fontSize: 30.0,
-                                fontWeight: FontWeight.bold),
-                            controller: _displayNameController,
-                            onChanged: (text) {
-                              setState(() {
-                                displayName = text;
-                              });
-                            },
-                            maxLines: 1,
-                            maxLength: 20,
-                            maxLengthEnforced: true,
-                          ),
-                        ),
-                        Container(
-                          padding: EdgeInsets.only(
-                            top: 5.0,
-                          ),
-                          child: Text(
-                            'About',
-                            style: TextStyle(
-                                fontSize: 20.0, fontWeight: FontWeight.bold),
-                          ),
-                          alignment: Alignment(-1.0, 0.0),
-                        ),
-                        Container(
-                          child: TextField(
-                            decoration:
-                                InputDecoration(border: OutlineInputBorder()),
-                            keyboardType: TextInputType.multiline,
-                            maxLength: 500,
-                            maxLengthEnforced: true,
-                            maxLines: null,
-                            onChanged: (text) {
-                              setState(() {
-                                aboutMe = text;
-                              });
-                            },
-                            controller: _aboutMeController,
-                          ),
-                        ),
-                        Row(
-                          children: <Widget>[
-                            Container(
-                              padding: EdgeInsets.only(top: 10.0),
-                              child: Text(
-                                'Interests',
-                                style: TextStyle(
-                                    fontSize: 20.0,
-                                    fontWeight: FontWeight.bold),
-                              ),
-                              alignment: Alignment(-1.0, 0.0),
-                            ),
-                            SizedBox(
-                              width: 20.0,
-                            ),
-                            Container(
-                              alignment: Alignment(1.0, 0.0),
-                              padding: EdgeInsets.only(top: 10.0),
-                              child: FlatButton(
-                                onPressed: () {
-                                  Navigator.push(context,
-                                      RouteFromRight(widget: Interests()));
-                                },
-                                color: Colors.orange,
-                                child: Text('Edit Interests'),
-                              ),
-                            ),
-                          ],
-                        ),
-                        Container(
-                            height: 75.0,
-                            child: ProfileInterestsList(
-                              interests: [], //List.of(usersSnapshot.data['interests'] ?? [])
-                            )
+    // return SafeArea(
+    //   child: ListView(
+    //     children: <Widget>[
+    //       Container(
+    //         child: Stack(
+    //           alignment: Alignment.topCenter,
+    //           children: <Widget>[
+    //             Column(
+    //               children: <Widget>[
+    //                 Container(
+    //                   height: 150.0,
+    //                 ),
+    //                 Card(
+    //                     child: Column(
+    //                   mainAxisSize: MainAxisSize.min,
+    //                   children: <Widget>[
+    //                     Container(
+    //                       alignment: Alignment(0.0, 1.0),
+    //                       height: 150.0,
+    //                       width: 500.0,
+    //                       color: Color(0x00333333),
+    //                       child: TextField(
+    //                         decoration: InputDecoration(
+    //                             prefixIcon: Icon(Icons.account_circle),
+    //                             helperText: "Your display name"),
+    //                         style: TextStyle(
+    //                             color: Colors.black,
+    //                             fontSize: 30.0,
+    //                             fontWeight: FontWeight.bold),
+    //                         controller: _displayNameController,
+    //                         onChanged: (text) {
+    //                           setState(() {
+    //                             displayName = text;
+    //                           });
+    //                         },
+    //                         maxLines: 1,
+    //                         maxLength: 20,
+    //                         maxLengthEnforced: true,
+    //                       ),
+    //                     ),
+    //                     Container(
+    //                       padding: EdgeInsets.only(
+    //                         top: 5.0,
+    //                       ),
+    //                       child: Text(
+    //                         'About',
+    //                         style: TextStyle(
+    //                             fontSize: 20.0, fontWeight: FontWeight.bold),
+    //                       ),
+    //                       alignment: Alignment(-1.0, 0.0),
+    //                     ),
+    //                     Container(
+    //                       child: TextField(
+    //                         decoration:
+    //                             InputDecoration(border: OutlineInputBorder()),
+    //                         keyboardType: TextInputType.multiline,
+    //                         maxLength: 500,
+    //                         maxLengthEnforced: true,
+    //                         maxLines: null,
+    //                         onChanged: (text) {
+    //                           setState(() {
+    //                             aboutMe = text;
+    //                           });
+    //                         },
+    //                         controller: _aboutMeController,
+    //                       ),
+    //                     ),
+    //                     Row(
+    //                       children: <Widget>[
+    //                         Container(
+    //                           padding: EdgeInsets.only(top: 10.0),
+    //                           child: Text(
+    //                             'Interests',
+    //                             style: TextStyle(
+    //                                 fontSize: 20.0,
+    //                                 fontWeight: FontWeight.bold),
+    //                           ),
+    //                           alignment: Alignment(-1.0, 0.0),
+    //                         ),
+    //                         SizedBox(
+    //                           width: 20.0,
+    //                         ),
+    //                         Container(
+    //                           alignment: Alignment(1.0, 0.0),
+    //                           padding: EdgeInsets.only(top: 10.0),
+    //                           child: FlatButton(
+    //                             onPressed: () {
+    //                               Navigator.push(context,
+    //                                   RouteFromRight(widget: Interests()));
+    //                             },
+    //                             color: Colors.orange,
+    //                             child: Text('Edit Interests'),
+    //                           ),
+    //                         ),
+    //                       ],
+    //                     ),
+    //                     Container(
+    //                         height: 75.0,
+    //                         child: ProfileInterestsList(
+    //                           interests: [], //List.of(usersSnapshot.data['interests'] ?? [])
+    //                         )
 
-                            // ListView.builder(
-                            //     scrollDirection: Axis.horizontal,
-                            //     itemCount: List.of(
-                            //             usersSnapshot.data['interests'])
-                            //         .length,
-                            //     itemBuilder:
-                            //         (BuildContext context, int index) {
-                            //       return ButtonBar(
-                            //         alignment: MainAxisAlignment.start,
-                            //         children: <Widget>[
-                            //           RaisedButton(
-                            //             onPressed: null,
-                            //             elevation: 5.0,
-                            //             child: Text('#' +
-                            //                 List.of(usersSnapshot
-                            //                         .data['interests'])
-                            //                     .elementAt(index)),
-                            //           ),
-                            //         ],
-                            //       );
-                            //     })
+    //                         // ListView.builder(
+    //                         //     scrollDirection: Axis.horizontal,
+    //                         //     itemCount: List.of(
+    //                         //             usersSnapshot.data['interests'])
+    //                         //         .length,
+    //                         //     itemBuilder:
+    //                         //         (BuildContext context, int index) {
+    //                         //       return ButtonBar(
+    //                         //         alignment: MainAxisAlignment.start,
+    //                         //         children: <Widget>[
+    //                         //           RaisedButton(
+    //                         //             onPressed: null,
+    //                         //             elevation: 5.0,
+    //                         //             child: Text('#' +
+    //                         //                 List.of(usersSnapshot
+    //                         //                         .data['interests'])
+    //                         //                     .elementAt(index)),
+    //                         //           ),
+    //                         //         ],
+    //                         //       );
+    //                         //     })
 
-                            ),
-                        SaveButton(
-                          uid: widget.uid,
-                          aboutMe: aboutMe,
-                          displayName: displayName,
-                        )
-                      ],
-                    )),
-                  ],
-                ),
-                UserImage(
-                  imgHeight: imgHeight,
-                  imgWidth: imgWidth,
-                  photoURL: usersSnapshot.data['photoURL'],
-                ),
-                // FadeInImage.assetNetwork(
-                //   placeholder: PROFILE_PHOTO_PLACEHOLDER_PATH,
-                //   fadeInCurve: SawTooth(1),
-                //   image: downloadURL,
-                //   width: 125.0,
-                //   height: 125.0,
-                //   fit: BoxFit.fitHeight,
-                // ),
-                Positioned(
-                  top: imgYPos + 120.0,
-                  // left: imgXPos + 150.0,
-                  child: FlatButton(
-                    color: Colors.orange,
-                    child: Text('Edit Picture'),
-                    onPressed: () {
-                      getImage();
-                    },
-                  ),
-                ),
-                CloseButton(),
-              ],
-            ),
-          ),
-        ],
-      ),
-    );
+    //                         ),
+    //                     SaveButton(
+    //                       uid: widget.uid,
+    //                       aboutMe: aboutMe,
+    //                       displayName: displayName,
+    //                     )
+    //                   ],
+    //                 )),
+    //               ],
+    //             ),
+    //             UserImage(
+    //               imgHeight: imgHeight,
+    //               imgWidth: imgWidth,
+    //               photoURL: usersSnapshot.data['photoURL'],
+    //             ),
+    //             // FadeInImage.assetNetwork(
+    //             //   placeholder: PROFILE_PHOTO_PLACEHOLDER_PATH,
+    //             //   fadeInCurve: SawTooth(1),
+    //             //   image: downloadURL,
+    //             //   width: 125.0,
+    //             //   height: 125.0,
+    //             //   fit: BoxFit.fitHeight,
+    //             // ),
+    //             Positioned(
+    //               top: imgYPos + 120.0,
+    //               // left: imgXPos + 150.0,
+    //               child: FlatButton(
+    //                 color: Colors.orange,
+    //                 child: Text('Edit Picture'),
+    //                 onPressed: () {
+    //                   getImage();
+    //                 },
+    //               ),
+    //             ),
+    //             CloseButton(),
+    //           ],
+    //         ),
+    //       ),
+    //     ],
+    //   ),
+    // );
   }
 
   // @override
