+++ lib/pages/profile/profile.dart
@@ -1,7 +1,5 @@
-import 'dart:async';
 import 'dart:core';
 
-import 'package:cloud_firestore/cloud_firestore.dart';
 import 'package:eatwithme/models/models.dart';
 import 'package:eatwithme/pages/chat/chat_room.dart';
 import 'package:eatwithme/services/db.dart';
@@ -14,7 +12,6 @@ import 'package:flutter/material.dart';
 import 'package:eatwithme/pages/interests/interests.dart';
 import 'package:eatwithme/main.dart';
 import 'package:eatwithme/widgets/loadingCircle.dart';
-import 'package:eatwithme/pages/auth/auth.dart';
 import 'package:provider/provider.dart';
 
 class ProfilePage extends StatefulWidget {
